Implemented a chunk-binned performance renderer with cached per-chunk splat ordering, adaptive LOD, visibleSplatIndices indirect drawing

+++ src/world/chunk-gpu-metadata.ts
@@ -26,12 +26,12 @@ export function packChunkGpuMetadata(chunks: readonly SplatChunk[]): ArrayBuffer
     view.setFloat32(base + 16, chunk.boundsMax[0], true);
     view.setFloat32(base + 20, chunk.boundsMax[1], true);
     view.setFloat32(base + 24, chunk.boundsMax[2], true);
-    view.setUint32(base + 28, chunk.lodLevel, true);
+    view.setFloat32(base + 28, chunk.lodLevel, true);
 
     view.setFloat32(base + 32, chunk.center[0], true);
     view.setFloat32(base + 36, chunk.center[1], true);
     view.setFloat32(base + 40, chunk.center[2], true);
-    view.setUint32(base + 44, chunk.splatCount, true);
+    view.setFloat32(base + 44, chunk.splatCount, true);
 
     view.setUint32(base + 48, chunk.id, true);
     view.setUint32(base + 52, chunk.splatStart, true);
@@ -41,4 +41,3 @@ export function packChunkGpuMetadata(chunks: readonly SplatChunk[]): ArrayBuffer
 
   return buffer;
 }
-
